move story prompts out
